# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Courtside TW API
  version: 1.0.0
  description: Contract-first REST API for public reading, reader library, editorial publication and optional provenance/SIWE boundaries.
servers:
  - url: https://api.courtside.tw
    description: Production API origin placeholder; deployment is a later task.
...
security:
  - bffSession: []
  - oidcBearer: []
paths:
...
  /api/v1/editor/media/{id}:
    get:
      tags:
        - Media
      operationId: getEditorMediaMetadata
      summary: Read media metadata
      description: Reads accessibility metadata and the latest rights record for an editor-owned media asset.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaMetadata"
              examples:
                valid:
                  value:
                    assetId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    version: 1
                    altText: Courtside action
                    state: READY
                    rights: null
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "429":
          $ref: "#/components/responses/Problem429"
      x-required-roles:
        - EDITOR
    patch:
      tags:
        - Media
      operationId: updateEditorMediaMetadata
      summary: Update media metadata and rights
      description: Persists alt text and a versioned rights record with an If-Match lock on the media asset.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaMetadataUpdate"
            examples:
              valid:
                value:
                  altText: Courtside action
                  rights:
                    rightsOwner: Courtside Studio
                    licenseName: Editorial license
                    allowedChannels:
                      - PUBLIC_WEB
                    territories:
                      - GLOBAL
                    validFrom: 2026-08-10T00:00:00Z
                    validUntil: 2027-08-10T00:00:00Z
                    credit: Courtside Studio
                    withdrawalTerms: Contact rights desk
                    status: VALID
      responses:
        "200":
          description: Metadata updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaMetadata"
              examples:
                valid:
                  value:
                    assetId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    version: 2
                    altText: Courtside action
                    state: READY
                    rights: null
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
        "429":
          $ref: "#/components/responses/Problem429"
      x-idempotent: false
      x-optimistic-lock: If-Match
      x-required-roles:
        - EDITOR
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC issuer/audience validated JWT for Spring Security.
    bffSession:
      type: apiKey
      in: cookie
      name: __Host-courtside_session
      description: Secure, HttpOnly, SameSite BFF session cookie.
  parameters:
...
    Id:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Optimistic-lock version or ETag. The server rejects stale values with 409 VERSION_CONFLICT.
      schema:
        type: string
        minLength: 1
        maxLength: 200
...
  headers:
    XRequestId:
      description: Trace/request identifier. Never contains tokens or article content.
      schema:
        type: string
        minLength: 1
        maxLength: 100
    ETag:
      description: Deterministic representation validator for conditional public reads.
      schema:
        type: string
        maxLength: 128
    RetryAfter:
      description: Seconds until the caller may retry.
      schema:
        type: integer
        minimum: 1
        maximum: 3600
  schemas:
    Uuid:
      type: string
      format: uuid
      description: Stable public identifier.
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - detail
        - instance
        - requestId
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 1
          maxLength: 1000
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          minLength: 1
          maxLength: 100
        code:
          type: string
          enum:
            - INVALID_REQUEST
            - AUTHENTICATION_REQUIRED
            - FORBIDDEN
            - RESOURCE_NOT_FOUND
            - VERSION_CONFLICT
            - RIGHTS_OR_CONTENT_GATE
            - RATE_LIMITED
        errors:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      additionalProperties: false
      required:
        - path
        - code
        - message
      properties:
        path:
          type: string
          minLength: 1
          maxLength: 200
        code:
          type: string
          minLength: 1
          maxLength: 80
        message:
          type: string
          minLength: 1
          maxLength: 500
...
    MediaMetadataUpdate:
      type: object
      additionalProperties: false
      required:
        - altText
      properties:
        altText:
          type: string
          minLength: 1
          maxLength: 1000
        rights:
          $ref: "#/components/schemas/MediaRightsMetadataInput"
    MediaMetadata:
      type: object
      additionalProperties: false
      required:
        - assetId
        - version
        - altText
        - state
        - rights
      properties:
        assetId:
          $ref: "#/components/schemas/Uuid"
        version:
          type: integer
          minimum: 0
        altText:
          oneOf:
            - type: string
              minLength: 1
              maxLength: 1000
            - type: "null"
        state:
          type: string
          enum:
            - PENDING
            - PROCESSING
            - READY
            - FAILED
            - REVOKED
        rights:
          oneOf:
            - $ref: "#/components/schemas/MediaRightsMetadata"
            - type: "null"
    MediaRightsMetadataInput:
      type: object
      additionalProperties: false
      required:
        - rightsOwner
        - licenseName
        - allowedChannels
        - territories
        - validFrom
        - validUntil
        - credit
        - withdrawalTerms
        - status
      properties:
        version:
          type: integer
          minimum: 0
        rightsOwner:
          type: string
          minLength: 1
          maxLength: 512
        licenseName:
          type: string
          minLength: 1
          maxLength: 512
        allowedChannels:
          type: array
          minItems: 1
          maxItems: 16
          items:
            type: string
            enum:
              - PUBLIC_WEB
              - READER_LIBRARY
              - OFFLINE
              - PROVENANCE
        territories:
          type: array
          minItems: 1
          maxItems: 16
          items:
            type: string
            pattern: ^[A-Za-z0-9_-]{1,64}$
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        credit:
          type: string
          minLength: 1
          maxLength: 1000
        withdrawalTerms:
          type: string
          minLength: 1
          maxLength: 2000
        status:
          type: string
          enum:
            - UNKNOWN
            - PENDING
            - VALID
            - EXPIRED
            - REVOKED
            - BLOCKED
    MediaRightsMetadata:
      allOf:
        - $ref: "#/components/schemas/MediaRightsMetadataInput"
        - type: object
          required:
            - id
            - version
          properties:
            id:
              $ref: "#/components/schemas/Uuid"
...
  responses:
    Problem400:
      description: Invalid request (400).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: INVALID_REQUEST example
              value:
                type: https://courtside.tw/problems/invalid_request
                title: Invalid request
                status: 400
                detail: The request is invalid.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: INVALID_REQUEST
      x-stable-error-code: INVALID_REQUEST
    Problem401:
      description: Authentication required (401).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: AUTHENTICATION_REQUIRED example
              value:
                type: https://courtside.tw/problems/authentication_required
                title: Authentication required
                status: 401
                detail: Authentication is required.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: AUTHENTICATION_REQUIRED
      x-stable-error-code: AUTHENTICATION_REQUIRED
    Problem403:
      description: Forbidden (403).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: FORBIDDEN example
              value:
                type: https://courtside.tw/problems/forbidden
                title: Forbidden
                status: 403
                detail: The actor is not allowed to perform this operation.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: FORBIDDEN
      x-stable-error-code: FORBIDDEN
    Problem404:
      description: Not found (404).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: RESOURCE_NOT_FOUND example
              value:
                type: https://courtside.tw/problems/resource_not_found
                title: Not found
                status: 404
                detail: The requested resource was not found.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: RESOURCE_NOT_FOUND
      x-stable-error-code: RESOURCE_NOT_FOUND
    Problem409:
      description: Conflict (409).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: VERSION_CONFLICT example
              value:
                type: https://courtside.tw/problems/version_conflict
                title: Conflict
                status: 409
                detail: The resource version is stale or the state conflicts.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: VERSION_CONFLICT
      x-stable-error-code: VERSION_CONFLICT
    Problem422:
      description: Unprocessable content (422).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: RIGHTS_OR_CONTENT_GATE example
              value:
                type: https://courtside.tw/problems/rights_or_content_gate
                title: Unprocessable content
                status: 422
                detail: The content or rights gate blocks this operation.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: RIGHTS_OR_CONTENT_GATE
      x-stable-error-code: RIGHTS_OR_CONTENT_GATE
    Problem429:
      description: Too many requests (429).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: RATE_LIMITED example
              value:
                type: https://courtside.tw/problems/rate_limited
                title: Too many requests
                status: 429
                detail: The request rate is limited.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: RATE_LIMITED
      x-stable-error-code: RATE_LIMITED
...